修改爬虫逻辑，完善项目结构

+++ admin/src/main/java/com/lsl/demo/first/common/dto/LevelDto.java
@@ -39,6 +39,6 @@ public class LevelDto {
     @Min(1)
     @NotNull
     @ApiModelProperty("电影评级 1 - 10")
-    private Integer level;
+    private Double level;
 
 }
